First experiment hardcoding settings and sheet URL
This version avoids saving/retrieving settings via scriptProperties
service, the idea is to reduce latency, this can be a third o quarter of a
second when the server response is fast.
So difference is not so great but at least it reduces internal complexity
and a little bit of latency.
Adminpage and configuration panel do not exist in this version, since
settings are found in the code itself.